restructure marketdataservice into multiple files for simpler organization
